+++ CHANGES.md
@@ -22,10 +22,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   a re-initialization of the optimizer (#369)
 - Support for scipy sparse CSR matrices as input (as, e.g., returned by sklearn's
   `CountVectorizer`); note that they are cast to dense matrices during batching
+- Helper functions to build command line interfaces with almost no
+  boilerplate, [example][1811191713] that shows usage
 
 [1810251445]: https://colab.research.google.com/github/dnouri/skorch/blob/master/notebooks/Basic_Usage.ipynb
 [1810261633]: https://colab.research.google.com/github/dnouri/skorch/blob/master/notebooks/Advanced_Usage.ipynb
 [1811011230]: https://colab.research.google.com/github/dnouri/skorch/blob/master/notebooks/MNIST.ipynb
+[1811191713]: https://github.com/dnouri/skorch/tree/master/examples/cli
 
 ### Changed